WIP moved the tests relating to _solve_hssp
