+++ optuna/_hypervolume/__init__.py
@@ -8,5 +8,6 @@ __all__ = [
     "BaseHypervolume",
     "_compute_2d",
     "_compute_2points_volume",
+    "_solve_hssp"
     "WFG",
 ]
